fix #59: for the second time: use Integer values for variable indices
This is the reversion of commit 65d5fbe0204ae6f247aa01c9512a21ec2f4acd4e
together with the fixing of the implementation of MinMax-mappend.

Now it works and the time is still acceptable (a 5% slowdown at most).

+++ src/Theory/Proof/CaseDistinctions.hs
@@ -296,7 +296,7 @@ precomputeCaseDistinctions ctxt typAsms =
 
     absFact (Fact tag ts) = (tag, length ts)
 
-    nMsgVars n = [ varTerm (LVar "t" LSortMsg i) | i <- [1..n] ]
+    nMsgVars n = [ varTerm (LVar "t" LSortMsg i) | i <- [1..fromIntegral n] ]
 
     someProtoGoal :: (FactTag, Int) -> LNFact
     someProtoGoal (tag, arity) = Fact tag (nMsgVars arity)
